fix(ci): line-limit guardrail — split oversized executor test shard
